els-ASAP auto commit file/tag libosinfo-1.0.0-1.el7.src.rpm

+++ libosinfo.spec
@@ -1,23 +1,12 @@
 # -*- rpm-spec -*-
 
-# Plugin isn't ready for real world use yet - it needs
-# a security audit at very least
-%define with_plugin 0
-
-%define with_gir 0
-
-%if 0%{?fedora} >= 15 || 0%{?rhel} >= 7
-%define with_gir 1
-%endif
-
 Summary: A library for managing OS information for virtualization
 Name: libosinfo
-Version: 0.3.0
-Release: 3%{?dist}%{?extra_release}
+Version: 1.0.0
+Release: 1%{?dist}%{?extra_release}
 License: LGPLv2+
 Group: Development/Libraries
 Source: https://fedorahosted.org/releases/l/i/%{name}/%{name}-%{version}.tar.gz
-
 BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
 URL: http://libosinfo.org/
 BuildRequires: intltool
@@ -29,24 +18,12 @@ BuildRequires: vala
 BuildRequires: vala-tools
 BuildRequires: libsoup-devel
 BuildRequires: /usr/bin/pod2man
-%if %{with_gir}
+BuildRequires: hwdata
 BuildRequires: gobject-introspection-devel
-%endif
+BuildRequires: osinfo-db
 Requires: hwdata
-
-# https://bugzilla.redhat.com/show_bug.cgi?id=1032524
-# Don't advertise unsigned drivers from unsecure locations
-Patch0: rm-unsigned-drivers-from-unsecure-locations.patch
-
-# https://bugzilla.redhat.com/show_bug.cgi?id=1069727
-# Support Windows 8.1 & 10 express install
-Patch1: enable-win8.1-win10-express-installation.patch
-
-# https://bugzilla.redhat.com/show_bug.cgi?id=1282919
-Patch2: libosinfo-0.3.0-add-ubuntu-16.04-iso.patch
-
-# https://bugzilla.redhat.com/show_bug.cgi?id=1375131
-Patch3: libosinfo-0.3.0-add-rhel-6.8-7.3-iso.patch
+Requires: osinfo-db
+Requires: osinfo-db-tools
 
 %description
 libosinfo is a library that allows virtualization provisioning tools to
@@ -82,19 +59,8 @@ This package provides the Vala bindings for libosinfo library.
 %prep
 %setup -q
 
-%patch0 -p1 -b .rm-unsigned-drivers-from-unsecure-locations
-%patch1 -p1 -b .enable-win8.1-win10-express-installation
-%patch2 -p1 -b .ubuntu-16.04-iso
-%patch3 -p1 -b .rhel-6.8--7.3-iso
-
 %build
-%if %{with_gir}
-%define gir_arg --enable-introspection=yes
-%else
-%define gir_arg --enable-introspection=no
-%endif
-
-%configure %{gir_arg} %{udev_arg} --enable-vala=yes --with-usb-ids-path=/usr/share/hwdata/usb.ids --with-pci-ids-path=/usr/share/hwdata/pci.ids
+%configure --enable-introspection=yes --enable-vala=yes
 %__make %{?_smp_mflags} V=1
 
 chmod a-x examples/*.js examples/*.py
@@ -108,7 +74,11 @@ rm -f %{buildroot}%{_libdir}/*.la
 %find_lang %{name}
 
 %check
-make check
+if ! make check
+then
+  cat test/test-suite.log || true
+  exit 1
+fi
 
 %clean
 rm -fr %{buildroot}
@@ -121,26 +91,13 @@ rm -fr %{buildroot}
 %defattr(-, root, root)
 %doc AUTHORS ChangeLog COPYING.LIB NEWS README
 %{_bindir}/osinfo-detect
-%{_bindir}/osinfo-db-validate
 %{_bindir}/osinfo-query
 %{_bindir}/osinfo-install-script
-%dir %{_datadir}/libosinfo/
-%dir %{_datadir}/libosinfo/db/
-%dir %{_datadir}/libosinfo/schemas/
-%{_datadir}/libosinfo/db/datamap
-%{_datadir}/libosinfo/db/device
-%{_datadir}/libosinfo/db/os
-%{_datadir}/libosinfo/db/platform
-%{_datadir}/libosinfo/db/install-script
-%{_datadir}/libosinfo/schemas/libosinfo.rng
-%{_mandir}/man1/osinfo-db-validate.1*
 %{_mandir}/man1/osinfo-detect.1*
 %{_mandir}/man1/osinfo-query.1*
 %{_mandir}/man1/osinfo-install-script.1*
 %{_libdir}/%{name}-1.0.so.*
-%if %{with_gir}
 %{_libdir}/girepository-1.0/Libosinfo-1.0.typelib
-%endif
 
 %files devel
 %defattr(-, root, root)
@@ -151,9 +108,7 @@ rm -fr %{buildroot}
 %dir %{_includedir}/%{name}-1.0/osinfo/
 %{_includedir}/%{name}-1.0/osinfo/*.h
 %{_libdir}/pkgconfig/%{name}-1.0.pc
-%if %{with_gir}
 %{_datadir}/gir-1.0/Libosinfo-1.0.gir
-%endif
 %{_datadir}/gtk-doc/html/Libosinfo
 
 %files vala
@@ -161,71 +116,48 @@ rm -fr %{buildroot}
 %{_datadir}/vala/vapi/libosinfo-1.0.vapi
 
 %changelog
-* Mon Sep 12 2016 David King <dking@redhat.com> - 0.3.0-3
-- Add data for RHEL 6.8 and 7.3 (#1375131)
+* Thu Feb 23 2017 Matthias Clasen <mclasen@redhat.com> 1.0.0-1
+- Rebase to 1.0.0
+  Resolves: rhbz#1387014
 
-* Thu Sep 08 2016 David King <dking@redhat.com> - 0.3.0-2
-- Add data for Ubuntu 16.04 ISOs (#1282919)
+* Fri Jul  1 2016 Daniel P. Berrange <berrange@redhat.com> 0.3.1-1
+- New upstream release 0.3.1
 
-* Mon May 16 2016 Zeeshan Ali <zeenix@redhat.com> - 0.3.0-1
-- Rebase libosinfo (#1069727).
-- Enable express installation for win8.1 & win10 (#1069727).
+* Thu Feb 04 2016 Fedora Release Engineering <releng@fedoraproject.org> - 0.3.0-2
+- Rebuilt for https://fedoraproject.org/wiki/Fedora_24_Mass_Rebuild
 
-* Fri Apr 15 2016 Zeeshan Ali <zeenix@redhat.com> - 0.2.12-4
-- Add patch that adds a lot of new data. (#1282919 #1257865 #1257867 #1257869).
+* Fri Jan  8 2016 Zeeshan Ali <zeenix@redhat.com> 0.3.0-1
+- New upstream release 0.3.0
 
-* Wed Aug 19 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.12-3
-- Bump relase to get the new build out. (#1197236).
+* Wed Jun 17 2015 Fedora Release Engineering <rel-eng@lists.fedoraproject.org> - 0.2.12-2
+- Rebuilt for https://fedoraproject.org/wiki/Fedora_23_Mass_Rebuild
 
-* Wed Aug 19 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.12-2
-- Add data about RHEL 7.1 & 7.2 (#1197236).
+* Thu May 28 2015 Zeeshan Ali <zeenix@redhat.com> 0.2.12-1
+- New upstream release 0.2.12
 
-* Thu May 28 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.12-1
-- Update to 0.2.12 (#1174552).
-- Remove now redundant patches.
+* Mon Sep 22 2014 Cole Robinson <crobinso@redhat.com> - 0.2.11-2
+- os: Add Fedora 21
 
-* Sat May 16 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.11-5
-- Merge RHELSA branch (#1197236).
+* Tue Aug 26 2014 Christophe Fergeau <cfergeau@redhat.com> 0.2.11-1
+- New upstream release 0.2.11
 
-* Mon Jan 26 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.11-4
-- Add/update F21 info (#1169415).
+* Sun Aug 17 2014 Fedora Release Engineering <rel-eng@lists.fedoraproject.org> - 0.2.9-4
+- Rebuilt for https://fedoraproject.org/wiki/Fedora_21_22_Mass_Rebuild
 
-* Mon Jan 26 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.11-3
-- Don't specify network interface in fedora script (#1169415).
+* Tue Jul 22 2014 Kalev Lember <kalevlember@gmail.com> - 0.2.9-3
+- Rebuilt for gobject-introspection 1.41.4
 
-* Thu Jan  8 2015 Zeeshan Ali <zeenix@redhat.com> - 0.2.11-2
-- Add rhel-5.9..11 and rhel-6.6 info (#1168199).
+* Sat Jun 07 2014 Fedora Release Engineering <rel-eng@lists.fedoraproject.org> - 0.2.9-2
+- Rebuilt for https://fedoraproject.org/wiki/Fedora_21_Mass_Rebuild
 
-* Wed Sep  3 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.11-1
-- Update to latest libosinfo release (#1128621).
+* Wed Dec 18 2013 Debarshi Ray <rishi@fedoraproject.org> - 0.2.9-1
+- New upstream release 0.2.9
 
-* Fri Feb 28 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-10
-- Rebuild with -fstack-protector-strong (#1070773).
+* Thu Nov 28 2013 Zeeshan Ali <zeenix@redhat.com> - 0.2.8-1
+- New upstream release 0.2.8
 
-* Fri Feb 28 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-9
-- Remove unsigned drivers from unsecure locations (#1032524).
-
-* Mon Feb 24 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-8
-- Disable unattended installation scripts for Windows 8.1 (#1067431).
-
-* Tue Feb 18 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-7
-- Support windows 8.1 (#1066042).
-
-* Fri Jan 24 2014 Daniel Mach <dmach@redhat.com> - 0.2.7-6
-- Mass rebuild 2014-01-24
-
-* Tue Jan  7 2014 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-5
-- fedora,win7: Improve & add data (#1047774).
-
-* Fri Dec 27 2013 Daniel Mach <dmach@redhat.com> - 0.2.7-4
-- Mass rebuild 2013-12-27
-
-* Thu Dec  5 2013 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-3
-- rhel: Add info on 6.5 and 7.0 (#1015055).
-- install-script,rhel: Fix for 'Server' & 7.0 (#1033007).
-
-* Fri Jul 19 2013 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-2
-- Don't install udev rule in RHEL7+ either (#983169).
+* Sat Aug 03 2013 Fedora Release Engineering <rel-eng@lists.fedoraproject.org> - 0.2.7-2
+- Rebuilt for https://fedoraproject.org/wiki/Fedora_20_Mass_Rebuild
 
 * Tue May 14 2013 Zeeshan Ali <zeenix@redhat.com> - 0.2.7-1
 - New upstream release 0.2.7
